Move exception setup to seperate source file
