+++ kernel/src/process_manager/mod.rs
@@ -13,6 +13,7 @@ pub mod memory_helper;
 pub mod allocation;
 pub mod memory_channels;
 pub mod outb;
+pub mod exception_handling;
 
 static MONITOR_INIT_STATE: ImmutAfterInitCell<bool> = ImmutAfterInitCell::new(false);
 const MONITOR_INIT_STATE_TRUE: bool = true;
